Move lib/constants.js to settings-template.js
