+++ lib/irc.js
@@ -1,6 +1,7 @@
 const irc = require('irc');
-const constants = require('./constants.js');
 const fs = require('fs');
+
+const constants = require('../settings.js');
 
 
 class IRC {
