+++ lib/mutes.js
@@ -1,4 +1,4 @@
-const constants = require('./constants.js');
+const constants = require('../settings.js');
 const IRC = require('./irc.js');
